+++ package.json
@@ -29,6 +29,7 @@
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.26.0",
+    "hint": "^7.1.3",
     "html-webpack-plugin": "^5.5.0",
     "jest": "^29.3.1",
     "style-loader": "^3.3.1",
