+++ src/modules/likeitems.js
@@ -1,6 +1,4 @@
 const url = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/kMoYqkIjxGGhhy0t9Ado/likes';
-const baseURL = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/';
-const appKey = 'kMoYqkIjxGGhhy0t9Ado';
 const getnumberofLikes = async () => {
   const cards = document.querySelectorAll('.card');
   const likeCount = document.querySelectorAll('.like-count');
